add spark master option and reorg common arg handling
