+++ mesoscope/commands/regions.py
@@ -5,6 +5,7 @@ from os.path import join, isfile, isdir, dirname, splitext, basename
 from skimage.io import imsave, imread
 from showit import image
 from extraction import load
+from .common import success, status, error, warn
 from ..models import compare as compareModels
 from ..models import overlay
 
@@ -56,15 +57,3 @@ def regions_command(input, output, image, compare, threshold, overwrite):
         warn('%s already exists and overwrite is false' % name)
 
     success('regions complete')
-
-def success(msg):
-    click.echo('[' + click.style('success', fg='green') + '] ' + msg)
-
-def status(msg):
-    click.echo('[' + click.style('convert', fg='blue') + '] ' + msg)
-
-def error(msg):
-    click.echo('[' + click.style('error', fg='red') + '] ' + msg)
-
-def warn(msg):
-    click.echo('[' + click.style('warn', fg='yellow') + '] ' + msg)
